renaming exception into returncode
